Added another pravus patch:
This patch changes the URL handling in the web portion of the Pastebot. I've added a simple URL decoder to allow '#' characters in the URL by using '%23'. The prefchan handling has also been changed to take this into account so that 'channel' and '#channel' both default the select box to '#channel', while '##channel' defaults to '##channel', etc.